licence info added to the metadata of the icons
I put the licence information in each file with XMP metadata.
I chose the CC BY-SA because it's an open licence compatible with GPLv3 and we need that in order to upload the plugin to the QGIS repository.
This ensures that anyone using these icons for anything and even modifying them will have to cite us as the original creators.